pyproject.toml
[project]
name = "mcp-pubmed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "mcp[cli]>=1.6.0",
    "pyyaml>=6.0.2",
]
